Refactor data module
Now we assume the CNN uses 'same' padding, so the image size does not reduce through the convolutional layers. So the HR image size is strictly LR image size multiplied by the scale factor. This assumption makes the code much cleaner.

Also the saving data to hdf5 part is removed. We are not actually using this functionality. And `load_set` does not take a long time to run.