PReLU still not working
We tried to use PReLU but haven't figured out a solution. The issue is that the trainable parameters in PReLU scale as x. So when we change the size of x, the whole network needs to be trained again, which is against our need for a network accepting arbitrary sized x.